Add basic unittest infrastucture(#120)
This commit adds the basic tools to run automated, headless unittests
against our javascript codebase. Doing so we now depend on ``npm`` for
javacript package management (only if you want to run tests as part of
the development setup).

Using ``npm`` we utilize multiple packages (plus their grunt plugins):
1. Grunt: As a task runner to handle execution and central
configuration.
2. QUnit as our JS unittest framework.
3. phantomjs for JS parsing. Because we need to make use of a central
``tests.html`` as point of entry for our tests (as far as I understand
to circumvent javascripts rediculous shortcommings with module imports)
we can not merely use node itself to run the tests.
Instead we use phantomjs to parse ``tests.html`` and take it from there.
4. ``babel``: Our codebase uses several ES6 features (e.g. ``let``,
``for-in``). Various of or tools however struggle with this and do not
yet support all of those features. In order to get around that we use
``babel`` to convert our original codebase into ES5 javascript and use
this result as imput for our tests. Please note that this workflow does
not provide full ES6 support (but it covers what we need right now). If
more is needed we may have to look at ``babel-polyfill``.
5. Istanbul for code coverage reports.
6. JSHint for code quality checks.

Most of this workflow is wrapped up in the new ``Gruntfile.js`` and new
make targets (``cleab-test``, ``coverage``, ``test``, ``test-all``).
It is worth noting that code quality (via JSHint) is currently not
checked automaticly as the codebase does not live up to it. This should
be handled in a seperate issue. You can run the checks manually with:
``grunt jshint``.

Writing tests:
* Read up the really simple API QUnit offers
* Open ``tests.js``
* Start writing tests.

!Please Note!
Currently only ``extension/stuff.js`` is actually covered by the test
framework. ``extension.js``, ``prefs.js`` and ``convenience.js`` use
GNOME-Shell's custom ``imports`` mechanism which we yet have not managed
to play along with the rest of our test environment.

Closes #120